backend/pyproject.toml
[tool.poetry]
name = "bloom-scroll-backend"
version = "0.1.0"
description = "Backend service for Bloom Scroll - A perspective-driven content aggregator"
authors = ["Bloom Scroll Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.104,<0.137"
uvicorn = {extras = ["standard"], version = ">=0.24,<0.47"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = ">=5.0.1,<8.0.0"
celery = "^5.3.4"
sentence-transformers = ">=2.2.2,<6.0.0"
transformers = "^4.35.0"
torch = "^2.1.0"
httpx = ">=0.25.1,<0.29.0"
beautifulsoup4 = "^4.12.2"
lxml = ">=4.9.3,<7.0.0"
pandas = ">=2.1.3,<4.0.0"
numpy = ">=1.26.2,<3.0.0"
scikit-learn = "^1.3.2"
pillow = ">=10.1,<13.0"
hdbscan = "^0.8.33"
pymilvus = "^2.3.3"
pgvector = "^0.2.3"
python-dotenv = "^1.0.0"
python-multipart = ">=0.0.6,<0.0.27"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = ">=23.2.1,<26.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3,<10.0.0"
pytest-asyncio = ">=0.21.1,<1.4.0"
pytest-cov = ">=4.1,<8.0"
black = ">=23.11,<27.0"
ruff = ">=0.1.6,<0.16.0"
mypy = "^1.7.0"
httpx = ">=0.25.1,<0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Alembic migrations are partially auto-generated; long lines mirror schema declarations.
"alembic/versions/*.py" = ["E501"]
# Top-level dev scripts use module-style constants and are not part of the package.
"test_story_001.py" = ["N806"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
